+++ packages/api-client/src/v2/features/modal/helpers/create-api-client-modal.ts
@@ -1,8 +1,7 @@
 import { type ModalState, useModal } from '@scalar/components'
-import type { ApiReferenceConfigurationRaw } from '@scalar/types/api-reference'
 import type { WorkspaceStore } from '@scalar/workspace-store/client'
 import { type WorkspaceEventBus, createWorkspaceEventBus } from '@scalar/workspace-store/events'
-import { type App, type MaybeRefOrGetter, computed, createApp, reactive } from 'vue'
+import { type App, computed, createApp, reactive } from 'vue'
 
 import {
   type DefaultEntities,
@@ -28,7 +27,7 @@ export type CreateApiClientModalOptions = {
   /** Api client plugins to include in the modal */
   plugins?: ClientPlugin[]
   /** Subset of the configuration options for the modal, if you want it to be reactive ensure its a ref */
-  options?: MaybeRefOrGetter<Pick<ApiReferenceConfigurationRaw, 'authentication' | 'proxyUrl'>>
+  options?: ModalProps['options']
 }
 
 export type ApiClientModal = {